+++ app/core/nice/quest.py
@@ -50,6 +50,7 @@ def get_nice_stage(
     return NiceStage(
         wave=raw_stage.wave,
         bgm=get_nice_bgm(region, raw_stage.bgmId),
+        fieldAis=raw_stage.script.get("aiFieldIds", []),
         enemies=enemies,
     )