+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - Quest enemy data from Rayshift to nice quest phase.
 - `phasesWithEnemies` to raw and nice quest.
+- `fieldAis` to nice stage.
 
 ## 5.51.0 - 2020-03-07
 ### Added